[fix] Fixed all linting errors (#10)
