+++ src/components/ConstructionComponent/ConstructionComponent.module.css
@@ -11,4 +11,5 @@
     gap: 10px;
     padding: 20px;
     justify-content: center;
+    padding-bottom: 50px;
 }
