+++ src/components/Footer/Footer.module.css
@@ -1,3 +1,13 @@
 .footer{
     display: block;
+    background-color: var(--background-color);
+    z-index: 10;
+    position: relative;
+    width: 100%;
+    height: 100%;
+}
+
+.footer h2{
+    margin: 0;
+    padding: 20px;
 }
