+++ src/components/SummaryComponent/SummaryComponent.module.css
@@ -4,6 +4,8 @@
   justify-content: center;
   gap: 2vw;
   margin: 3vw;
+  z-index: 1;
+  position: relative;
 }
 
 .personal_info_container {
