+++ src/context/ThemeContext.ts
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+interface ThemeContextProps {
+  theme: "light" | "dark";
+  toggleTheme: () => void;
+}
+
+export const ThemeContext = createContext<ThemeContextProps | undefined>(
+  undefined,
+);
