+++ src/types/image.ts
@@ -6,4 +6,11 @@ export type ImageGridProps = {
 export interface imagesArray {
     image: string,
     span: number,
+    alt: string,
 }
+
+export type Experience = {
+  title: string;
+  images: imagesArray[];
+  description: string;
+};
